# apps/war-room/lib/sag-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Slack Agent Gateway",
    "description": "Gateway service for multiple AI agents on Slack",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/slack/oauth/callback": {
      "get": {
        "tags": [
          "slack-agent-gateway"
        ],
        "summary": "Oauth Callback",
        "description": "Handle OAuth callback from Slack after app installation.\n\nThis endpoint is called by Slack when an admin completes the OAuth flow.\nSlack redirects the admin's browser here with an authorization code.\n\nNo API key auth because this is a browser redirect from Slack.\n\nArgs:\n    code: Authorization code from Slack (exchanged for bot token).\n    state: State parameter (request_id) for CSRF protection.\n    error: Error code if the user denied authorization.\n\nReturns:\n    HTML page showing success or error message.",
        "operationId": "oauth_callback_api_v1_slack_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/slack-agent-gateway/slack/oauth/callback": {
      "get": {
        "tags": [
          "slack-agent-gateway"
        ],
        "summary": "Oauth Callback",
        "description": "Handle OAuth callback from Slack after app installation.\n\nThis endpoint is called by Slack when an admin completes the OAuth flow.\nSlack redirects the admin's browser here with an authorization code.\n\nNo API key auth because this is a browser redirect from Slack.\n\nArgs:\n    code: Authorization code from Slack (exchanged for bot token).\n    state: State parameter (request_id) for CSRF protection.\n    error: Error code if the user denied authorization.\n\nReturns:\n    HTML page showing success or error message.",
        "operationId": "oauth_callback_slack_agent_gateway_slack_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}